Add files via upload
A number of teachers at a large school district are absent and need to be covered. These vacancies can be either filled by other regular teachers or by substitute teachers, who are paid at a lower rate. Therefore, the district prefers to employ subs to fill the demand. However, the subs get paid for being in the pool, even if they do not end up working. Using empirical data of teacher absences on any given weekday, the goal is to find the optimal number of substitutes to minimize the costs, particularly of using regular teachers and hiring too many subs. 

With the table of the cumulative percentages of teacher absences, we can create a slope for each data interval. Then, the inverse linear spline uses these slopes and the endpoints of each interval. This is a function that outputs the number of teachers demanded for any input between 0 and 1 (percentile of data). With a way to simulate demand, we can calculate the expected cost of any sized pool of subs by running thousands of simulations, which will show the different costs at various levels of demand.